% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demspaces-package.R
\docType{package}
\name{demspaces-package}
\alias{demspaces}
\alias{demspaces-package}
\title{demspaces: Democratic Spaces Barometer}
\description{
Helper package with model tooling and Shiny apps for the Democratic Spaces Barometer project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vdeminstitute.github.io/demspaces}
  \item \url{https://github.com/vdeminstitute/demspaces}
  \item Report bugs at \url{https://github.com/vdeminstitute/demspaces/issues}
}

}
\author{
\strong{Maintainer}: Andreas Beger \email{adbeger@gmail.com} (\href{https://orcid.org/0000-0003-1883-3169}{ORCID})

}
\keyword{internal}
