% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg.R
\name{ds_logistic_reg}
\alias{ds_logistic_reg}
\title{Logistic regression}
\usage{
ds_logistic_reg(space, data)
}
\arguments{
\item{space}{(\code{character(1)}) \cr
Name of the V-Dem indicator for a democratic space}

\item{data}{(\code{\link[base:data.frame]{base::data.frame()}})
A data frame or similar object}
}
\description{
Demonstrates the interface. Otherwise only difference is that it internally
normalizes input data before fitting and predicting.
}
