% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode-coder.R
\name{episode_coder}
\alias{episode_coder}
\alias{changes_one_country}
\title{Episode coder}
\usage{
episode_coder(x, cp, direction = c("up", "down"), type = c("mod", "orig"))

changes_one_country(x, cp)
}
\arguments{
\item{x}{vector containing values for one of the space indicators, for
\emph{one country} only.}

\item{cp}{a single value for the space cutpoint}

\item{direction}{code opening ("up") or closing ("down") events?}

\item{type}{use the original single-year event coding ("orig"), or the ERT-lite coding ("mod")?}
}
\value{
\code{episode_coder} returns a binary (integer) vector with \code{length(x)}
marking opening or closing events in \code{x}

\code{changes_one_country} returns a character vector with \code{length(x)}
and values "same", "up", "down"
}
\description{
Identifies opening/closing episodes in a stream of space indicator values.
}
