% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-reg-pca.R
\name{logistic_reg_pca}
\alias{logistic_reg_pca}
\title{Logistic regression with PCA}
\usage{
logistic_reg_pca(x, y)
}
\arguments{
\item{x}{Data frame with features.}

\item{y}{Binary vector indicating outcome event.}
}
\description{
Standardized interface for logistic regression with PCA feature variable
dimensionality reduction.
}
\examples{
credit_data <- recipes::credit_data

#mdl <- logistic_reg(credit_data[, setdiff(colnames(credit_data), "Status")],
#                     credit_data$Status)

}
