% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome-explorer-app.R
\name{outcome_explorer_app}
\alias{outcome_explorer_app}
\title{Outcome Explorer Shiny App}
\usage{
outcome_explorer_app()
}
\description{
For exploring the democratic spaces outcomes in different countries over time.
}
