% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spaces_for_app}
\alias{spaces_for_app}
\alias{spaces_for_app_orig,}
\alias{spaces_for_app_mod}
\alias{spaces_for_app_orig}
\title{Original and new outcomes}
\format{
An object of class \code{data.frame} with 11758 rows and 33 columns.

An object of class \code{data.frame} with 11758 rows and 33 columns.
}
\usage{
spaces_for_app_orig

spaces_for_app_mod
}
\description{
Original (<=v12) democratic spaces outcomes and new ERT-lite outcomes
(>=v12.1). This data is used by the \code{\link[=outcome_explorer_app]{outcome_explorer_app()}}.
}
\keyword{datasets}
